requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
cryptography>=41.0.0
pycryptodome>=3.19.0
selenium>=4.15.0
fake-useragent>=1.4.0
aiohttp>=3.9.0
tenacity>=8.2.0
